requirements.txt
boto3==1.24.65
mysql-connector-python==8.0.30
fastapi==0.82.0
pyyaml==6.0
from-root==1.0.2
uvicorn[standard]
python-multipart==0.0.5
requests==2.28.1
tqdm==4.64.1
pillow==9.2.0
dnspython
pymongo==4.2.0
typing==3.7.4.3
python-dotenv
-e .